Update skyblock_tips.snbt
Remove the "/skyblock create" command from the rewards and replace with 100 xp. The "/skyblock create" command is very disruptive to games in progress, leading to hours of bridge building or a restore if the reward is clicked and cheats are off.  Fixes issue #307 https://github.com/AllTheMods/atm6-sky/issues/372